requirements.txt
streamlit
numpy
pillow
opencv-python
tensorflow
matplotlib
streamlit-drawable-canvas
